convertCynapseMqttDataToInfluxDbData/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
influxdb==5.3.1
msgpack==1.0.2
paho-mqtt==1.5.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3